Changes without Optional 2

+++ src/main/java/com/example/burgerordersystem/Model/Menu.java
@@ -16,7 +16,7 @@ public class Menu {
 
     private int id;
     private String name;
-    private BigDecimal price;
+    private int price;
 
     private MainDish mainDish;
     private SideDish sideDish;
